Moved API_URL into config file inside fairmint folder
