+++ src/fairmint/reflectStakeholder.js
@@ -1,6 +1,6 @@
 import axios from "axios";
 import get from "lodash/get";
-import { API_URL } from "../chain-operations/utils";
+import { API_URL } from "./config";
 
 export const reflectStakeholder = async ({ issuerId, stakeholder }) => {
     console.log("Reflecting Stakeholder into fairmint...");
